update: page urls for documents.
